+++ packages/console/__tests__/Console.test.ts
@@ -78,4 +78,11 @@ describe("Console tests", () => {
         expect((cons as any).writeStdout).toHaveBeenCalledTimes(numberOfCalls);
         expect((cons as any).writeStderr).toHaveBeenCalledTimes(numberOfCalls);
     });
+
+    it("Should default to the same prefix as log4js", () => {
+        const cons = new Console();
+        jest.spyOn(Date.prototype, "getTimezoneOffset").mockReturnValueOnce(0);
+        jest.spyOn(Date, "now").mockReturnValueOnce(45296789);
+        expect((cons as any).buildPrefix("test")).toBe("[1970/01/01 12:34:56.789] [test] ");
+    });
 });